+++ src/lib/components/ButtonInline.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
-	import { concatClasses } from '$lib/utils/utils.js';
-	import { btnPrimary } from '$lib/utils/styles.js';
-	import type { Icon as LucideIcon } from 'lucide-svelte';
+	import { concatClasses } from '$lib/utils/utils'
+	import { btnPrimary } from '$lib/utils/styles'
+	import type { Icon as LucideIcon } from 'lucide-svelte'
 
 	type Props = {
 		Icon: LucideIcon
@@ -11,29 +11,19 @@
 		start?: boolean
 		end?: boolean
 	}
-	let { Icon, onclick, href, external, start = false, end = false }: Props = $props();
+	let { Icon, onclick, href, external, start = false, end = false }: Props = $props()
 
-	const sideClass = start
-		? 'rounded-l-md'
-		: end
-		? 'rounded-r-md'
-		: '';
-	const classes = concatClasses(btnPrimary, sideClass, 'w-12 min-w-12');
-	const aProps = external
-		? { target: '_blank', rel: 'noopener noreferrer' }
-		: {};
+	const sideClass = start ? 'rounded-l-md' : end ? 'rounded-r-md' : ''
+	const classes = concatClasses(btnPrimary, sideClass, 'w-12 min-w-12')
+	const aProps = external ? { target: '_blank', rel: 'noopener noreferrer' } : {}
 </script>
 
 {#if href}
-	<a
-		class={classes}
-		href={href}
-		{...aProps}
-	>
+	<a class={classes} {href} {...aProps}>
 		<Icon size={16} />
 	</a>
 {:else}
-	<button class={classes} onclick={onclick}>
+	<button class={classes} {onclick}>
 		<Icon size={16} />
 	</button>
 {/if}