+++ src/lib/components/LinkChip.svelte
@@ -4,7 +4,7 @@
 	import { isValidLink } from '$lib/utils/utils'
 	import { deleteLink } from '$lib/firestore/links'
 	import sessionStore from '$lib/stores/session.svelte'
-	import ButtonInline from '$lib/components/ButtonInline.svelte';
+	import ButtonInline from '$lib/components/ButtonInline.svelte'
 
 	type Props = {
 		link: Firestore.Doc<Firestore.Link>
@@ -30,24 +30,12 @@
 </script>
 
 <div class="flex">
-	<ButtonInline
-		Icon={OctagonX}
-		onclick={removeLink}
-		start
-	/>
+	<ButtonInline Icon={OctagonX} onclick={removeLink} start />
 	<div class="border-surface-800 max-h-19 min-h-9.5 flex-1 overflow-x-auto border px-4 py-1.5">
 		{link.url}
 	</div>
 	{#if isValidLink(link.url)}
-		<ButtonInline
-			Icon={ExternalLink}
-			href={link.url}
-			external
-		/>
+		<ButtonInline Icon={ExternalLink} href={link.url} external />
 	{/if}
-	<ButtonInline
-		Icon={Copy}
-		onclick={() => setClipboard(link.url)}
-		end
-	/>
+	<ButtonInline Icon={Copy} onclick={() => setClipboard(link.url)} end />
 </div>