+++ src/lib/components/SignUpForm.svelte
@@ -38,7 +38,7 @@
 			placeholder="Email"
 			autocomplete="email"
 			bind:value={inputEmail}
-			constraints={{ required: true }}
+			attr={{ required: true }}
 		>
 			{#snippet icon()}
 				<AtSign size={16} />
@@ -51,7 +51,7 @@
 			autocomplete="new-password"
 			canReveal
 			bind:value={inputPwd}
-			constraints={{ minlength: 6, maxlength: 64, required: true }}
+			attr={{ minlength: 6, maxlength: 64, required: true }}
 		>
 			{#snippet icon()}
 				<RectangleEllipsis size={16} />
@@ -64,7 +64,7 @@
 			autocomplete="new-password"
 			canReveal
 			bind:value={inputConfirmPwd}
-			constraints={{ minlength: 6, maxlength: 64, required: true }}
+			attr={{ minlength: 6, maxlength: 64, required: true }}
 		>
 			{#snippet icon()}
 				<RectangleEllipsis size={16} />